Move ParallelBus to its own module
