+++ ports/mimxrt10xx/common-hal/paralleldisplay/ParallelBus.c
@@ -3,7 +3,7 @@
  *
  * The MIT License (MIT)
  *
- * Copyright (c) 2019 Lucian Copeland for Adafruit Industries
+ * Copyright (c) 2020 Artur Pacholec
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,7 +24,7 @@
  * THE SOFTWARE.
  */
 
-#include "shared-bindings/displayio/ParallelBus.h"
+#include "shared-bindings/paralleldisplay/ParallelBus.h"
 
 #include <stdint.h>
 
@@ -33,35 +33,35 @@
 #include "shared-bindings/digitalio/DigitalInOut.h"
 #include "shared-bindings/microcontroller/__init__.h"
 
-void common_hal_displayio_parallelbus_construct(displayio_parallelbus_obj_t *self,
+void common_hal_paralleldisplay_parallelbus_construct(paralleldisplay_parallelbus_obj_t *self,
     const mcu_pin_obj_t *data0, const mcu_pin_obj_t *command, const mcu_pin_obj_t *chip_select,
     const mcu_pin_obj_t *write, const mcu_pin_obj_t *read, const mcu_pin_obj_t *reset, uint32_t frequency) {
 
     mp_raise_NotImplementedError(translate("ParallelBus not yet supported"));
 }
 
-void common_hal_displayio_parallelbus_deinit(displayio_parallelbus_obj_t *self) {
+void common_hal_paralleldisplay_parallelbus_deinit(paralleldisplay_parallelbus_obj_t *self) {
 
 }
 
-bool common_hal_displayio_parallelbus_reset(mp_obj_t obj) {
+bool common_hal_paralleldisplay_parallelbus_reset(mp_obj_t obj) {
     return false;
 }
 
-bool common_hal_displayio_parallelbus_bus_free(mp_obj_t obj) {
+bool common_hal_paralleldisplay_parallelbus_bus_free(mp_obj_t obj) {
     return false;
 }
 
-bool common_hal_displayio_parallelbus_begin_transaction(mp_obj_t obj) {
+bool common_hal_paralleldisplay_parallelbus_begin_transaction(mp_obj_t obj) {
 
     return false;
 }
 
-void common_hal_displayio_parallelbus_send(mp_obj_t obj, display_byte_type_t byte_type,
+void common_hal_paralleldisplay_parallelbus_send(mp_obj_t obj, display_byte_type_t byte_type,
     display_chip_select_behavior_t chip_select, const uint8_t *data, uint32_t data_length) {
 
 }
 
-void common_hal_displayio_parallelbus_end_transaction(mp_obj_t obj) {
+void common_hal_paralleldisplay_parallelbus_end_transaction(mp_obj_t obj) {
 
 }
