+++ py/circuitpy_mpconfig.h
@@ -330,6 +330,7 @@ extern const struct _mp_obj_module_t canio_module;
 // CIRCUITPY_COUNTIO uses MP_REGISTER_MODULE
 // CIRCUITPY_DIGITALIO uses MP_REGISTER_MODULE
 // CIRCUITPY_DISPLAYIO uses MP_REGISTER_MODULE
+// CIRCUITPY_PARALLELDISPLAY uses MP_REGISTER_MODULE
 // CIRCUITPY_TERMINALIO uses MP_REGISTER_MODULE
 // CIRCUITPY_FONTIO uses MP_REGISTER_MODULE
