+++ py/circuitpy_mpconfig.mk
@@ -154,6 +154,13 @@ CFLAGS += -DCIRCUITPY_COUNTIO=$(CIRCUITPY_COUNTIO)
 CIRCUITPY_DISPLAYIO ?= $(CIRCUITPY_FULL_BUILD)
 CFLAGS += -DCIRCUITPY_DISPLAYIO=$(CIRCUITPY_DISPLAYIO)
 
+ifeq ($(CIRCUITPY_DISPLAYIO),1)
+CIRCUITPY_PARALLELDISPLAY ?= $(CIRCUITPY_FULL_BUILD)
+else
+CIRCUITPY_PARALLELDISPLAY = 0
+endif
+CFLAGS += -DCIRCUITPY_PARALLELDISPLAY=$(CIRCUITPY_PARALLELDISPLAY)
+
 # bitmaptools and framebufferio rely on displayio
 ifeq ($(CIRCUITPY_DISPLAYIO),1)
 CIRCUITPY_BITMAPTOOLS ?= $(CIRCUITPY_FULL_BUILD)
