+++ shared-bindings/paralleldisplay/ParallelBus.h
@@ -24,31 +24,31 @@
  * THE SOFTWARE.
  */
 
-#ifndef MICROPY_INCLUDED_SHARED_BINDINGS_DISPLAYBUSIO_PARALLELBUS_H
-#define MICROPY_INCLUDED_SHARED_BINDINGS_DISPLAYBUSIO_PARALLELBUS_H
+#ifndef MICROPY_INCLUDED_SHARED_BINDINGS_DISPLAYBUSIO_PARALLELDISPLAY_H
+#define MICROPY_INCLUDED_SHARED_BINDINGS_DISPLAYBUSIO_PARALLELDISPLAY_H
 
-#include "common-hal/displayio/ParallelBus.h"
+#include "common-hal/paralleldisplay/ParallelBus.h"
 
 #include "common-hal/microcontroller/Pin.h"
 #include "shared-bindings/displayio/__init__.h"
 #include "shared-module/displayio/Group.h"
 
-extern const mp_obj_type_t displayio_parallelbus_type;
+extern const mp_obj_type_t paralleldisplay_parallelbus_type;
 
-void common_hal_displayio_parallelbus_construct(displayio_parallelbus_obj_t *self,
+void common_hal_paralleldisplay_parallelbus_construct(paralleldisplay_parallelbus_obj_t *self,
     const mcu_pin_obj_t *data0, const mcu_pin_obj_t *command, const mcu_pin_obj_t *chip_select,
     const mcu_pin_obj_t *write, const mcu_pin_obj_t *read, const mcu_pin_obj_t *reset, uint32_t frequency);
 
-void common_hal_displayio_parallelbus_deinit(displayio_parallelbus_obj_t *self);
+void common_hal_paralleldisplay_parallelbus_deinit(paralleldisplay_parallelbus_obj_t *self);
 
-bool common_hal_displayio_parallelbus_reset(mp_obj_t self);
-bool common_hal_displayio_parallelbus_bus_free(mp_obj_t self);
+bool common_hal_paralleldisplay_parallelbus_reset(mp_obj_t self);
+bool common_hal_paralleldisplay_parallelbus_bus_free(mp_obj_t self);
 
-bool common_hal_displayio_parallelbus_begin_transaction(mp_obj_t self);
+bool common_hal_paralleldisplay_parallelbus_begin_transaction(mp_obj_t self);
 
-void common_hal_displayio_parallelbus_send(mp_obj_t self, display_byte_type_t byte_type,
+void common_hal_paralleldisplay_parallelbus_send(mp_obj_t self, display_byte_type_t byte_type,
     display_chip_select_behavior_t chip_select, const uint8_t *data, uint32_t data_length);
 
-void common_hal_displayio_parallelbus_end_transaction(mp_obj_t self);
+void common_hal_paralleldisplay_parallelbus_end_transaction(mp_obj_t self);
 
-#endif // MICROPY_INCLUDED_SHARED_BINDINGS_DISPLAYBUSIO_PARALLELBUS_H
+#endif // MICROPY_INCLUDED_SHARED_BINDINGS_DISPLAYBUSIO_PARALLELDISPLAY_H
